# Conflicts:
#	app/src/main/ic_launcher-playstore.png
#	app/src/main/res/mipmap-hdpi/ic_launcher.png
#	app/src/main/res/mipmap-hdpi/ic_launcher_foreground.png
#	app/src/main/res/mipmap-hdpi/ic_launcher_round.png
#	app/src/main/res/mipmap-mdpi/ic_launcher.png
#	app/src/main/res/mipmap-mdpi/ic_launcher_foreground.png
#	app/src/main/res/mipmap-mdpi/ic_launcher_round.png
#	app/src/main/res/mipmap-xhdpi/ic_launcher.png
#	app/src/main/res/mipmap-xhdpi/ic_launcher_foreground.png
#	app/src/main/res/mipmap-xhdpi/ic_launcher_round.png
#	app/src/main/res/mipmap-xxhdpi/ic_launcher.png
#	app/src/main/res/mipmap-xxhdpi/ic_launcher_foreground.png
#	app/src/main/res/mipmap-xxhdpi/ic_launcher_round.png
#	app/src/main/res/mipmap-xxxhdpi/ic_launcher.png
#	app/src/main/res/mipmap-xxxhdpi/ic_launcher_foreground.png
#	app/src/main/res/mipmap-xxxhdpi/ic_launcher_round.png
#	app/src/main/res/values/ic_launcher_background.xml

+++ app/src/main/java/com/gov/sidesa/data/letter/models/WidgetResponse.kt
@@ -14,10 +14,10 @@ import kotlinx.parcelize.Parcelize
 data class WidgetResponse(
     @SerializedName("id_type_surat_document")
     @Expose
-    val documentTypeId: Int? = null,
+    val documentTypeId: Long? = null,
     @SerializedName("id_type_surat")
     @Expose
-    val letterTypeId: Int? = null,
+    val letterTypeId: Long? = null,
     @SerializedName("name")
     @Expose
     val name: String = "",
